refactor: create submodules for terminals and berths in dropins (#106)
