+++ fis/dropins/graph.py
@@ -11,164 +11,6 @@ logger = logging.getLogger(__name__)
 geod = Geod(ellps="WGS84")
 
 
-def generate_terminal_graph_features(terminals: List[Dict]) -> List[Dict]:
-    """
-    Generates node and edge features for terminals.
-    Each terminal gets a node and an 'access' edge connecting it to the
-    fairway junction node created during splicing.
-    """
-    features = []
-    for term in terminals:
-        raw_id = term.get("id", term.get("Id"))
-        if raw_id is None:
-            logger.warning("Skipping terminal without 'id'/'Id': %s", term)
-            continue
-        tid = utils.stringify_id(raw_id)
-        conn_wkt = term.get("connection_geometry")
-        if not conn_wkt:
-            # Terminal was not mapped to a section or section was not spliced
-            continue
-
-        conn_pt = wkt.loads(conn_wkt)
-        term_geom = term.get("geometry")
-        if not term_geom:
-            continue
-        term_pt = wkt.loads(term_geom) if isinstance(term_geom, str) else term_geom
-        if term_pt.geom_type != "Point":
-            term_pt = term_pt.centroid
-
-        # 1. Connection node on the fairway (where the split happened)
-        conn_id = f"terminal_{tid}_connection"
-        features.append(
-            {
-                "type": "Feature",
-                "geometry": mapping(conn_pt),
-                "properties": {
-                    "id": conn_id,
-                    "feature_type": "node",
-                    "node_type": "junction",
-                    "node_id": conn_id,
-                },
-            }
-        )
-
-        # 2. Terminal node itself
-        features.append(
-            {
-                "type": "Feature",
-                "geometry": mapping(term_pt),
-                "properties": {
-                    "id": f"terminal_{tid}",
-                    "feature_type": "node",
-                    "node_type": "terminal",
-                    "node_id": f"terminal_{tid}",
-                    "name": term.get("Name"),
-                    "terminal_id": tid,
-                    "isrs_id": term.get("IsrsId"),
-                },
-            }
-        )
-
-        # 3. Access edge
-        # LineString from connection point on fairway to terminal point
-        access_line = LineString([conn_pt, term_pt])
-        features.append(
-            {
-                "type": "Feature",
-                "geometry": mapping(access_line),
-                "properties": {
-                    "id": f"terminal_access_{tid}",
-                    "feature_type": "fairway_segment",
-                    "segment_type": "terminal_access",
-                    "terminal_id": tid,
-                    "source_node": conn_id,
-                    "target_node": f"terminal_{tid}",
-                    "length_m": geod.geometry_length(access_line),
-                },
-            }
-        )
-
-    return features
-
-
-def generate_berth_graph_features(berths: List[Dict]) -> List[Dict]:
-    """
-    Generates node and edge features for berths.
-    Each berth gets a node and an 'access' edge connecting it to the
-    fairway junction node created during splicing.
-    """
-    features = []
-    for berth in berths:
-        raw_id = berth.get("id", berth.get("Id"))
-        if raw_id is None:
-            logger.warning("Skipping berth without 'id'/'Id': %s", berth)
-            continue
-        bid = utils.stringify_id(raw_id)
-        conn_wkt = berth.get("connection_geometry")
-        if not conn_wkt:
-            continue
-
-        conn_pt = wkt.loads(conn_wkt)
-        berth_geom = berth.get("geometry")
-        if not berth_geom:
-            continue
-        berth_pt = wkt.loads(berth_geom) if isinstance(berth_geom, str) else berth_geom
-        if berth_pt.geom_type != "Point":
-            berth_pt = berth_pt.centroid
-
-        # 1. Connection node on the fairway
-        conn_id = f"berth_{bid}_connection"
-        features.append(
-            {
-                "type": "Feature",
-                "geometry": mapping(conn_pt),
-                "properties": {
-                    "id": conn_id,
-                    "feature_type": "node",
-                    "node_type": "junction",
-                    "node_id": conn_id,
-                },
-            }
-        )
-
-        # 2. Berth node itself
-        features.append(
-            {
-                "type": "Feature",
-                "geometry": mapping(berth_pt),
-                "properties": {
-                    "id": f"berth_{bid}",
-                    "feature_type": "node",
-                    "node_type": "berth",
-                    "node_id": f"berth_{bid}",
-                    "name": berth.get("Name"),
-                    "berth_id": bid,
-                    "isrs_id": berth.get("IsrsId"),
-                },
-            }
-        )
-
-        # 3. Access edge
-        access_line = LineString([conn_pt, berth_pt])
-        features.append(
-            {
-                "type": "Feature",
-                "geometry": mapping(access_line),
-                "properties": {
-                    "id": f"berth_access_{bid}",
-                    "feature_type": "fairway_segment",
-                    "segment_type": "berth_access",
-                    "berth_id": bid,
-                    "source_node": conn_id,
-                    "target_node": f"berth_{bid}",
-                    "length_m": geod.geometry_length(access_line),
-                },
-            }
-        )
-
-    return features
-
-
 def generate_simplified_passages(
     complexes: List[Dict], structure_type: str
 ) -> List[Dict]:
